Headers cleanup
