+++ libmamba/src/api/list.cpp
@@ -4,8 +4,6 @@
 //
 // The full license is in the file LICENSE, distributed with this software.
 
-#include <regex>
-
 #include "mamba/api/list.hpp"
 #include "mamba/core/channel.hpp"
 #include "mamba/api/configuration.hpp"
